usercontroller edit MyPageController remove, UserCharacterController remove
