--- src/main/java/com/scriptopia/demo/controller/MyPageController.java
@@ -1,35 +1,0 @@
-package com.scriptopia.demo.controller;
-
-import com.scriptopia.demo.dto.history.HistoryPageResponse;
-import com.scriptopia.demo.service.GameSessionService;
-import com.scriptopia.demo.service.HistoryService;
-import com.scriptopia.demo.service.SharedGameService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-import java.util.UUID;
-
-@RestController
-@RequiredArgsConstructor
-public class MyPageController {
-    private final HistoryService historyService;
-    private final SharedGameService sharedGameService;
-    private final GameSessionService gameSessionService;
-
-
-    /*
-    유저 -> 사용자 게임 기록 조회
-     */
-    @GetMapping("/my-page/history")
-    public ResponseEntity<List<HistoryPageResponse>> getHistory(@RequestParam(required = false) UUID lastId,
-                                                                @RequestParam(defaultValue = "10") int size,
-                                                                Authentication authentication) {
-        Long userId = Long.valueOf(authentication.getName());
-
-        return historyService.fetchMyHistory(userId, lastId, size);
-    }
-
-}
